requirements.txt
fastapi
uvicorn[standard]
asyncpg
uuid6
python-dotenv
